+++ app/MacroCHallengeApp/MacroCHallengeApp/Implementations/Controllers/SchoolsViewControllerImplementation.swift
@@ -54,14 +54,3 @@ class SchoolsViewControllerImplementation: UIViewController, SchoolsViewControll
     }
 }
 
-extension UITabBarController {
-	open override var childForStatusBarStyle: UIViewController? {
-		return selectedViewController?.childForStatusBarStyle ?? selectedViewController
-	}
-}
-
-extension UINavigationController {
-	open override var childForStatusBarStyle: UIViewController? {
-		return topViewController?.childForStatusBarStyle ?? topViewController
-	}
-}
