+++ app/MacroCHallengeApp/MacroCHallengeApp/Util/Extensions.swift
@@ -1,0 +1,21 @@
+//
+//  Extensions.swift
+//  MacroCHallengeApp
+//
+//  Created by João Pedro de Amorim on 07/10/20.
+//
+
+import Foundation
+import UIKit
+
+extension UITabBarController {
+    open override var childForStatusBarStyle: UIViewController? {
+        return selectedViewController?.childForStatusBarStyle ?? selectedViewController
+    }
+}
+
+extension UINavigationController {
+    open override var childForStatusBarStyle: UIViewController? {
+        return topViewController?.childForStatusBarStyle ?? topViewController
+    }
+}
